BasePaletteの定義を抜き出した
